Big app restructuring: add instance folder, update create_app, and use docker-compose.
