+++ config/settings.py
@@ -1,0 +1,9 @@
+# development
+ENV = 'development'
+DEBUG = True
+HOST = '0.0.0.0'
+PORT = 8000
+
+# database
+SQLALCHEMY_TRACK_MODIFICATIONS = False
+SQLALCHEMY_DATABASE_URI = 'your_database_credentials'
