+++ docker-compose.yaml
@@ -1,0 +1,16 @@
+version: '2'
+
+services:
+    api:
+        build: .
+        command: >
+          gunicorn -b 0.0.0.0:8000
+            --access-logfile -
+            --reload
+            "hotel_api.app:create_app()"
+        env_file:
+          - '.env'
+        volumes:
+          - '.:/hotel_api'
+        ports:
+          - '8000:8000'
