+++ hotel_api/common/utils.py
