+++ hotel_api/extensions.py
@@ -1,0 +1,5 @@
+from flask_restful import Api
+from flask_sqlalchemy import SQLAlchemy
+
+api = Api()
+db = SQLAlchemy()
