+++ instance/settings.py_production_example
@@ -1,0 +1,1 @@
+SQLALCHEMY_DATABASE_URI = 'your_database_credentials'
